Runtime unification Phase 3d: MapHeader bridge layout
The two map headers diverged on dispatch storage: VM held u32 bytecode
indices for user-defined Hash/Eq, AOT held C function pointers. The
bridge layout in `roxy_map_header` now carries BOTH — function pointers
for AOT dispatch, indices for VM dispatch — so a single struct serves
both runtimes during the migration. Phase 5 will collapse this by
routing VM dispatch through a thread-local trampoline and dropping the
index fields.

- `roxy_map_header` gains `uint32_t hash_fn_index; uint32_t eq_fn_index;`
  after the function-pointer block. AOT writes `UINT32_MAX` for both;
  VM writes the bytecode indices and leaves `hash_fn`/`eq_fn` null.
- Field offsets for `key_kind` / `key_slot_count` / `key_is_inline` /
  `value_slot_count` / `value_is_inline` are unchanged, so
  INDEX_GET_MAP / INDEX_SET_MAP at interpreter.cpp:1591-1645 continue
  to read the same byte offsets.
- `MapKeyKind` enum class re-anchored to the C `ROXY_MAP_KEY_*`
  constants; `using MapHeader = roxy_map_header;` in vm/map.hpp.
- VM-side switch statements on `header->key_kind` cast through
  `MapKeyKind` (the field is now `uint8_t`); 3 sites updated.
- `roxy_map_alloc` (AOT) initializes the index fields to `UINT32_MAX`;
  vm/map.cpp `map_alloc` initializes `hash_fn`/`eq_fn` to null.

Verified: 26 Maps cases, 12 Traits cases (struct keys with custom
Hash/Eq), 68 C-backend cases — all pass.

+++ src/roxy/vm/natives.cpp
@@ -550,7 +550,7 @@ static void native_map_init(RoxyVM* vm, u8 dst, u8 argc, u8 first_arg) {
     // First arg after self is key_kind
     if (argc >= 2) {
         i32 kind_val = static_cast<i32>(regs[first_arg + 1]);
-        header->key_kind = key_kind_from_i32(kind_val);
+        header->key_kind = static_cast<u8>(key_kind_from_i32(kind_val));
     }
 
     // Second arg after self is optional capacity
